refactor: rename Data interface to IDataListData and move into ts/types.tsx, use in IAboutSectionContent && ICompanyData
Renamed Data interface to IDataListData to clarify the purpose and origin. Moved the Interface into the ts/types.tsx file to reuse the interface in the IAboutSectionContent.skills.frontend && backend key and ICompanyData.skills. This adjustment was necessary to ensure the same data structure in the component which provides the data (AboutSection) and the component which utilises that data (DataListCard).