style: Increase z-index of navbarHeader in Navbar.module.scss
This adjustment ensures that the navbarHeader stays on top of all other elements, preventing any animated content from overlapping it. The z-index of navbarHeader has been increased to 3 to achieve this. This modification impacts the layout up to the "l" breakpoint, as the affected element is not displayed beyond this breakpoint.